test: Exclude

+++ nest-app/src/user/user.service.ts
@@ -1,11 +1,15 @@
-import {
-  Injectable,
-  NotFoundException,
-  ConflictException,
-} from "@nestjs/common";
+import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { User } from "./entities/user.entity";
+
+import { toDto } from "../util/transformer";
+import { UserDto } from "./dto/user.dto";
+import { CreateUserDto } from "./dto/create-user.dto";
+import {
+  UserNotFoundException,
+  UsernameAlreadyExistsException,
+} from "../global/exception/user.exception";
 
 @Injectable()
 export class UserService {
@@ -15,19 +19,20 @@ export class UserService {
   ) {}
 
   // 创建用户，添加查重逻辑
-  async create(user: User): Promise<User> {
+  async create(user: CreateUserDto): Promise<UserDto> {
     // 查重：根据 username 查找是否已有相同用户
-    const existingUser = await this.userRepository.findOne({
-      where: { username: user.username }, // 只根据 username 查重
+    const existingUser = await this.findOneByCondition({
+      username: user.username,
     });
 
     if (existingUser) {
-      throw new ConflictException("Username already exists");
+      throw new UsernameAlreadyExistsException(user.username);
     }
 
     // 如果没有重复，则创建新用户
     const newUser = this.userRepository.create(user);
-    return this.userRepository.save(newUser);
+    const updatedUser = await this.userRepository.save(newUser);
+    return toDto(updatedUser, UserDto);
   }
 
   async findAll({
@@ -37,46 +42,50 @@ export class UserService {
     pageNo?: number;
     pageSize?: number;
   }) {
-    // 计算跳过的条数
-    const skip = (pageNo - 1) * pageSize;
-
     // 使用 findAndCount 来获取分页数据和总数
     const [items, total] = await this.userRepository.findAndCount({
-      select: ["id", "username"], // 选择返回的字段，可以根据需求进行调整
-      skip, // 跳过的条数
+      skip: (pageNo - 1) * pageSize, // 跳过的条数
       take: pageSize, // 每页的条数
     });
 
     // 返回分页结果
     return {
-      items,
+      items: toDto(items, UserDto),
       total,
       pageNo,
       pageSize,
-      totalPages: Math.ceil(total / pageSize), // 计算总页数
     };
   }
 
-  async findOneById(id: number): Promise<User> {
-    const user = await this.userRepository.findOne({
-      where: { id },
-      relations: [], // 根据需要加入关联查询
+  findOneByCondition(where: Record<string, any>): Promise<User | null> {
+    return this.userRepository.findOne({
+      where,
+      relations: [], // 这里可以根据需要配置关联查询
     });
+  }
+
+  async findOneById(id: number): Promise<UserDto> {
+    const user = await this.findOneByCondition({ id });
     if (!user) {
-      throw new NotFoundException(`User with ID ${id} not found`);
+      throw new UserNotFoundException(id);
     }
-    return user;
+    return toDto(user, UserDto);
   }
 
-  async update(id: number, user: User) {
-    const oldUser = await this.findOneById(id);
-    this.userRepository.merge(oldUser, user);
-    // 联合模型更新，需要使用save
-    return this.userRepository.save(oldUser);
+  async update(id: number, userData: Partial<User>): Promise<UserDto> {
+    const user = await this.findOneByCondition({ id });
+    if (!user) throw new UserNotFoundException(id);
+
+    this.userRepository.merge(user, userData);
+    const updatedUser = await this.userRepository.save(user);
+    return toDto(updatedUser, UserDto);
   }
 
-  async delete(id: number) {
-    const user = await this.findOneById(id);
-    return this.userRepository.remove(user);
+  async delete(id: number): Promise<UserDto> {
+    const user = await this.findOneByCondition({ id });
+    if (!user) throw new UserNotFoundException(id);
+
+    const deleteUser = await this.userRepository.remove(user);
+    return toDto(deleteUser, UserDto);
   }
 }
